Sweep Environment
Created some folders for the swepes. Also plotted the first sweep for alpha_R and scenario 2. Created a parameter file such that we only need to make changes there. Its still to be applied in the model (model does not yet take the input from the parameter file - so far only for plotting)